Fixed #5 Projects can no longer be deleted after their goal is reached; Closes #6 IllegalArguments will now provide a 400 Error instead of the entire Exception

+++ src/main/java/de/clerkvest/api/entity/project/ProjectService.java
@@ -82,6 +82,9 @@ public class ProjectService implements IService<Project> {
 
     @Override
     public void delete(Project project) {
+        if (project.isReached()) {
+            throw new ViolatedConstraintException("Funded Projects can't be deleted");
+        }
         repository.delete(project);
     }
 }
